[mlir][linalg] Fix the naming of grouped conv2d
The name `conv_2d_ngchw_fgchw` suggests it takes a kernel in FGCHW format, but it actually expected the kernel to be in GFCHW format, and Torch-MLIR is using this op on the assumption that it takes GFCHW kernel.
https://github.com/llvm/torch-mlir/blob/0cf9ee340bc3264922e03b0df002e41c69184702/lib/Conversion/TorchToLinalg/Linear.cpp#L826

This change renames `conv_2d_ngchw_fgchw` to `conv_2d_ngchw_gfchw` and fixes the shape map.
Note that the behavior of this op remains the same as before.

Also adds a test against static tensors to make sure the data format is correct.

Differential Revision: https://reviews.llvm.org/D150064